requirements.txt
torch
torchvision
gradio
captum
Pillow
matplotlib
numpy
fastapi
uvicorn
sqlalchemy
pydantic-settings
python-multipart